requirements.txt
flask
gunicorn
beautifulsoup4
line-bot-sdk
flask_caching